Revert doctype decleration
Remove doctype tag. Due to the fact that webapp is not rendered
correctly in standars mode, the doctype should be added along
with the appropriate html/css `quirks mode ->standards mode`
related changes. We temporary revert the addition of the tag.

The doctype declaration will be recovered once we investigate
the standard mode rendering issues.